+++ KWD/XMOS/HID/src/HIDKeywordDetector.cpp
@@ -19,6 +19,7 @@
 #include <fcntl.h>
 #include <unistd.h>
 #include <chrono>
+#include <dirent.h>
 
 #include <AVSCommon/Utils/Logger/Logger.h>
 
@@ -39,38 +40,14 @@ static const std::string TAG("HIDKeywordDetector");
  */
 #define LX(event) alexaClientSDK::avsCommon::utils::logger::LogEntry(TAG, event)
 
-/// Keyword string
-static const std::string KEYWORD_STRING = "alexa";
-
-/// The number of hertz per kilohertz.
-static const size_t HERTZ_PER_KILOHERTZ = 1000;
-
-/// The timeout to use for read calls to the SharedDataStream.
-const std::chrono::milliseconds TIMEOUT_FOR_READ_CALLS = std::chrono::milliseconds(1000);
-
-/// The HID KW compatible AVS sample rate of 16 kHz.
-static const unsigned int HID_COMPATIBLE_SAMPLE_RATE = 16000;
-
-/// The HID KW compatible bits per sample of 16.
-static const unsigned int HID_COMPATIBLE_SAMPLE_SIZE_IN_BITS = 16;
-
-/// The HID KW compatible number of channels, which is 1.
-static const unsigned int HID_COMPATIBLE_NUM_CHANNELS = 1;
-
-/// The HID KW compatible audio encoding of LPCM.
-static const avsCommon::utils::AudioFormat::Encoding HID_COMPATIBLE_ENCODING =
-    avsCommon::utils::AudioFormat::Encoding::LPCM;
-
-/// The HID KW compatible endianness which is little endian.
-static const avsCommon::utils::AudioFormat::Endianness HID_COMPATIBLE_ENDIANNESS =
-    avsCommon::utils::AudioFormat::Endianness::LITTLE;
-
 /// HID keycode to monitor:
 static const char * HID_KEY_CODE = "KEY_T";
 
-/// HID device path
+/// HID device directory path
+static const std::string HID_DEVICE_DIR_PATH =  "/dev/input/";
 
-static const char * HID_DEVICE_PATH =  "/dev/input/event0";
+/// HID device name
+static const std::string HID_DEVICE_NAME = "XMOS XVF3615 Voice Processor Keyboard";
 
 /// USB Product ID of XMOS device
 static const int USB_VENDOR_ID = 0x20B1;
@@ -87,59 +64,52 @@ static const int CONTROL_RESOURCE_ID = 0xE0;
 /// The command ID of the XMOS control command.
 static const int CONTROL_CMD_ID = 0xAF;
 
-/// The lenght of the payload of the XMOS control command
+/// The length of the payload of the XMOS control command
 /// one control byte plus 3 uint64_t values
 static const int CONTROL_CMD_PAYLOAD_LEN = 25;
 
-/**
- * Read a specific index from the payload of the USB control message
- *
- * @param payload The data returned via control message
- * @param start_index The index in the payload to start reading from
- * @return value stored in payload
- */
-static uint64_t readIndex(uint8_t* payload, int start_index) {
-    uint64_t u64value = 0;
-    // convert array of bytes into uint64_t value
-    memcpy(&u64value, &payload[start_index], sizeof(uint64_t));
-    // swap bytes of uint64_t value
-    u64value = __bswap_64(u64value);
-    return u64value;
-}
-
-/**
- * Search for an USB device, open the connection and return the correct handlers
- *
- * @param evdev The device handler necessary for reading HID events
- * @param devh  The device handler necessary for sending control commands
- * @return 0 if device is found and handlers correctly set
- */
-static uint8_t openUSBDevice(libevdev** evdev, libusb_device_handle** devh) {
+bool HIDKeywordDetector::openDevice() {
     int rc = 1;
     libusb_device **devs = NULL;
     libusb_device *dev = NULL;
 
-    ACSDK_INFO(LX("openUSBDeviceOngoing")
-               .d("HIDDevicePath", HID_DEVICE_PATH)
+    ACSDK_INFO(LX("openDeviceOngoing")
+               .d("HIDDeviceName", HID_DEVICE_NAME)
                .d("USBVendorID", USB_VENDOR_ID)
                .d("USBProductID", USB_PRODUCT_ID));
 
     // Find USB device for reading HID events
-    int fd;
-    fd = open(HID_DEVICE_PATH, O_RDONLY|O_NONBLOCK);
-    rc = libevdev_new_from_fd(fd, evdev);
-    if (rc < 0) {
-        ACSDK_ERROR(LX("openUSBDeviceFailed")
-                        .d("reason", "initialiseLibevdevFailed")
-                        .d("error", strerror(-rc)));
-        return -1;
-    }
+    int fd = -1;
+    DIR *dir;
+    struct dirent *entry;
 
+    //search all files in directory
+    dir = opendir(HID_DEVICE_DIR_PATH.c_str()); 
+    if (dir) {
+        while ((entry = readdir(dir)) != NULL) {
+            std::string file_path(entry->d_name);
+            fd = open((HID_DEVICE_DIR_PATH+file_path).c_str(), O_RDONLY|O_NONBLOCK);
+
+            // Do not check if command is successful, as the entries reported by readdir() are not all devices
+            rc = libevdev_new_from_fd(fd, &m_evdev);
+            if (!rc) {
+                if (libevdev_get_name(m_evdev)==HID_DEVICE_NAME) {
+                    ACSDK_INFO(LX("openDeviceSuccess").d("reason", "Found HID device").d("path", HID_DEVICE_DIR_PATH+file_path));
+                    break;
+                }
+            }
+            fd = -1;
+        }
+        closedir(dir); //close directory
+    }
+    if (fd==-1) {
+        ACSDK_ERROR(LX("openDeviceFailed").d("reason", "HidDeviceNotFound"));
+    }
     // Find USB device for sending control commands
     int ret = libusb_init(NULL);
     if (ret < 0) {
-        ACSDK_ERROR(LX("openUSBDeviceFailed").d("reason", "initialiseLibUsbFailed"));
-        return -1;
+        ACSDK_ERROR(LX("openDeviceFailed").d("reason", "initialiseLibUsbFailed"));
+        return false;
     }
 
     int num_dev = libusb_get_device_list(NULL, &devs);
@@ -154,63 +124,17 @@ static uint8_t openUSBDevice(libevdev** evdev, libusb_device_handle** devh) {
     }
 
     if (dev == NULL) {
-        ACSDK_ERROR(LX("openUSBDeviceFailed").d("reason", "UsbDeviceNotFound"));
-        return -1;
+        ACSDK_ERROR(LX("openDeviceFailed").d("reason", "UsbDeviceNotFound"));
+        return false;
     }
 
-    if (libusb_open(dev, devh) < 0) {
-        ACSDK_ERROR(LX("openUSBDeviceFailed").d("reason", "UsbDeviceNotOpened"));
-        return -1;
+    if (libusb_open(dev, &m_devh) < 0) {
+        ACSDK_ERROR(LX("openDeviceFailed").d("reason", "UsbDeviceNotOpened"));
+        return false;
     }
 
     libusb_free_device_list(devs, 1);
-    ACSDK_INFO(LX("openUSBDeviceSuccess").d("reason", "UsbDeviceOpened"));
-    return 0;
-}
-
-
-/**
- * Checks to see if an @c avsCommon::utils::AudioFormat is compatible with HID KW.
- *
- * @param audioFormat The audio format to check.
- * @return @c true if the audio format is compatible with HID KW and @c false otherwise.
- */
-static bool isAudioFormatCompatibleWithHIDKW(avsCommon::utils::AudioFormat audioFormat) {
-    if (HID_COMPATIBLE_ENCODING != audioFormat.encoding) {
-        ACSDK_ERROR(LX("isAudioFormatCompatibleWithHIDKWFailed")
-                        .d("reason", "incompatibleEncoding")
-                        .d("gpiowwEncoding", HID_COMPATIBLE_ENCODING)
-                        .d("encoding", audioFormat.encoding));
-        return false;
-    }
-    if (HID_COMPATIBLE_ENDIANNESS != audioFormat.endianness) {
-        ACSDK_ERROR(LX("isAudioFormatCompatibleWithHIDKWFailed")
-                        .d("reason", "incompatibleEndianess")
-                        .d("gpiowwEndianness", HID_COMPATIBLE_ENDIANNESS)
-                        .d("endianness", audioFormat.endianness));
-        return false;
-    }
-    if (HID_COMPATIBLE_SAMPLE_RATE != audioFormat.sampleRateHz) {
-        ACSDK_ERROR(LX("isAudioFormatCompatibleWithHIDKWFailed")
-                        .d("reason", "incompatibleSampleRate")
-                        .d("gpiowwSampleRate", HID_COMPATIBLE_SAMPLE_RATE)
-                        .d("sampleRate", audioFormat.sampleRateHz));
-        return false;
-    }
-    if (HID_COMPATIBLE_SAMPLE_SIZE_IN_BITS != audioFormat.sampleSizeInBits) {
-        ACSDK_ERROR(LX("isAudioFormatCompatibleWithHIDKWFailed")
-                        .d("reason", "incompatibleSampleSizeInBits")
-                        .d("gpiowwSampleSizeInBits", HID_COMPATIBLE_SAMPLE_SIZE_IN_BITS)
-                        .d("sampleSizeInBits", audioFormat.sampleSizeInBits));
-        return false;
-    }
-    if (HID_COMPATIBLE_NUM_CHANNELS != audioFormat.numChannels) {
-        ACSDK_ERROR(LX("isAudioFormatCompatibleWithHIDKWFailed")
-                        .d("reason", "incompatibleNumChannels")
-                        .d("gpiowwNumChannels", HID_COMPATIBLE_NUM_CHANNELS)
-                        .d("numChannels", audioFormat.numChannels));
-        return false;
-    }
+    ACSDK_INFO(LX("openDeviceSuccess").d("reason", "UsbDeviceOpened"));
     return true;
 }
 
@@ -232,10 +156,6 @@ std::unique_ptr<HIDKeywordDetector> HIDKeywordDetector::create(
         return nullptr;
     }
 
-    if (!isAudioFormatCompatibleWithHIDKW(audioFormat)) {
-        return nullptr;
-    }
-
     std::unique_ptr<HIDKeywordDetector> detector(new HIDKeywordDetector(
         stream, keyWordObservers, keyWordDetectorStateObservers, audioFormat));
 
@@ -253,53 +173,19 @@ HIDKeywordDetector::HIDKeywordDetector(
     std::unordered_set<std::shared_ptr<KeyWordDetectorStateObserverInterface>> keyWordDetectorStateObservers,
     avsCommon::utils::AudioFormat audioFormat,
     std::chrono::milliseconds msToPushPerIteration) :
-        AbstractKeywordDetector(keyWordObservers, keyWordDetectorStateObservers),
-        m_stream{stream},
-        m_maxSamplesPerPush((audioFormat.sampleRateHz / HERTZ_PER_KILOHERTZ) * msToPushPerIteration.count()) {
+        XMOSKeywordDetector(stream, keyWordObservers, keyWordDetectorStateObservers, audioFormat, msToPushPerIteration) {
 }
 
 HIDKeywordDetector::~HIDKeywordDetector() {
-    m_isShuttingDown = true;
-    if (m_detectionThread.joinable())
-        m_detectionThread.join();
-    if (m_readAudioThread.joinable())
-        m_readAudioThread.join();
-
 }
+
 
 bool HIDKeywordDetector::init() {
-    m_streamReader = m_stream->createReader(AudioInputStream::Reader::Policy::BLOCKING);
-    if (!m_streamReader) {
-        ACSDK_ERROR(LX("initFailed").d("reason", "createStreamReaderFailed"));
-        return false;
+    if (XMOSKeywordDetector::init()) {
+        m_detectionThread = std::thread(&HIDKeywordDetector::detectionLoop, this);
+        return true;
     }
-
-    if (openUSBDevice(&m_evdev, &m_devh) != 0) {
-        return false;
-    }
-
-    m_isShuttingDown = false;
-    m_readAudioThread = std::thread(&HIDKeywordDetector::readAudioLoop, this);
-    m_detectionThread = std::thread(&HIDKeywordDetector::detectionLoop, this);
-    return true;
-}
-
-void HIDKeywordDetector::readAudioLoop() {
-    std::vector<int16_t> audioDataToPush(m_maxSamplesPerPush);
-    bool didErrorOccur = false;
-    while (!m_isShuttingDown) {
-        readFromStream(
-            m_streamReader,
-            m_stream,
-            audioDataToPush.data(),
-            audioDataToPush.size(),
-            TIMEOUT_FOR_READ_CALLS,
-            &didErrorOccur);
-        if (didErrorOccur) {
-            m_isShuttingDown = true;
-        }
-
-    }
+    return false;
 }
 
 void HIDKeywordDetector::detectionLoop() {
@@ -313,6 +199,7 @@ void HIDKeywordDetector::detectionLoop() {
         auto currentIndex = m_streamReader->tell();
         struct input_event ev;
         rc = libevdev_next_event(m_evdev, LIBEVDEV_READ_FLAG_NORMAL, &ev);
+
         // wait for HID_KEY_CODE true event
         if (rc == 0 && strcmp(libevdev_event_type_get_name(ev.type), "EV_KEY")==0 && \
             strcmp(libevdev_event_code_get_name(ev.type, ev.code), HID_KEY_CODE)==0 && \