+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Feature enhancements, updates, and resolved issues from all releases are availab
 **XMOS-only change** - version 3:
 
     * disable building of Sensory KWD engine for all the products
+    * refactor and tidy up XMOS keyword detector classes
 
 ### Version 1.24.0 - June 1 2021
 Feature enhancements, updates, and resolved issues from all releases are available on the [Amazon developer portal](https://developer.amazon.com/docs/alexa/avs-device-sdk/release-notes.html)